refactor(react): reorganize useAuthenticateOAuth hooks by platform scope
