+++ packages/react/src/index.ts
@@ -17,7 +17,6 @@ export type {
   ZeroDevWalletConnectorParams,
 } from './connector.js'
 export { zeroDevWallet } from './connector.js'
-export { useAuthenticateOAuth } from './hooks/useAuthenticateOAuth.js'
 export { useAuthenticators } from './hooks/useAuthenticators.js'
 export { useExportPrivateKey } from './hooks/useExportPrivateKey.js'
 export { useExportWallet } from './hooks/useExportWallet.js'
@@ -37,3 +36,4 @@ export {
   OAUTH_PROVIDERS,
   verifyGoogleLoginUrl,
 } from './utils/verifyGoogleLoginUrl.js'
+export { useAuthenticateOAuth } from './web/useAuthenticateOAuth.js'
